net/nfp: support flow API for CoreNIC firmware
Add the flow validate/create/destroy/flush API of nfp PMD with CoreNIC
firmware.

The flow create API construct a control cmsg and send it to
firmware, then add this flow to the hash table.

The flow destroy API construct a control cmsg and send it to
firmware, then delete this flow from the hash table.

The flow flush API just iterate the flows in hash table and
call the flow destroy API.

Signed-off-by: Chaoyong He <chaoyong.he@corigine.com>
Reviewed-by: Long Wu <long.wu@corigine.com>
Reviewed-by: Peng Zhang <peng.zhang@corigine.com>